Complete User Authentication
This change finalizes user authentication as things stand at this time.
The API uses token-based authentication, where web clients will receive
a token with an expiration date (14 days in the future) and mobile clients
will receive a token with an infinite lifespan. Of course, clients will need
to provide the API with valid credentials to receive any type of token.

The following changes were made to implement this feature:

* Created ExpiringTokenAuthentication class, in authentication.py, to be
  used as main form of backend auth
* Created custom token class ScrumScrumUserToken in models.py
* Removed references to JSON Web Tokens as they will not be used as of now
* Refactored ObtainExpiringAuthToken in views.py to support use of custom token
* Updated default auth classes in REST_FRAMEWORK config inside settings.py
* Added some comments to improve readability
* Changed/added migrations, since the backend database changed with addition
  of custom token.