Complete Email Verification
This change implements email verification for new users. When a user registers,
their account will be set to inactive by default until the supplied email
is verified. Emails are verified by sending an encoded hyperlink to the given
email. When the link is clicked or pasted into the browser, the activate/
endpoint is called and the user's ID and activation key are extracted and
decoded from the URL. If the decoded key is found in the
'scrum_scrum_api_activationkey' table and has not expired, the user's account
is activated.

NOTE: Only JSON is returned from this endpoint... It would be a good plan to
      eventually have DJango render a very basic HTML page that can handle the
      server response, since this endpoint will be called directly from an
      email client rather than a mobile or web app.

Changes made:

* Added settings in django.conf.settings corresponding to activation keys.
* models.py
    - ScrumScrumUser now defaults to inactive
    - Added ActivationKey
    - Added ArchivedActivationKey
* Added url route to api/activate/<encoded_user_id>/<encoded_activation_key>
* views.py
    - Added activate(), as an api_view, which attempts to activate an account
* Added new file activation.py, which contains activation-specific methods
* Added templates/activation_email.html as an email template to be sent to
  a new user with an activation link.